cashflow-ai/backend/requirements.txt
python-dotenv==1.0.0
flask==3.0.0
flask-cors==4.0.0
